Move resolvers to new package
